chore: modified perceptron model to use SIMD matrix instead of normal matrices to perform efficiently

+++ src/test/java/neuralnetworks/perceptron/PerceptronTest.java
@@ -12,7 +12,7 @@ import java.util.Random;
 public class PerceptronTest {
     // Use a fixed random seed for reproducibility in tests that involve random data generation
     private static final long RANDOM_SEED = 42L;
-    private final double learningRate = 0.1;
+    private final float learningRate = 0.1f;
 
     @Nested
     @DisplayName("Logic Gate Tests")
@@ -20,14 +20,14 @@ public class PerceptronTest {
         @Test
         @DisplayName("Perceptron learns 3-input AND gate")
         void testANDGateTraining() {
-            double[][] inputs = threeInputTruthTable();
+            float[][] inputs = threeInputTruthTable();
             int[] outputs = {0, 0, 0, 0, 0, 0, 0, 1}; // Only (1, 1, 1) should output 1 for AND gate
 
             Perceptron perceptron = new Perceptron(inputs, outputs, learningRate);
             perceptron.train(100);
 
             for (int i = 0; i < inputs.length; i++) {
-                double actualOutput = perceptron.predict(inputs[i]);
+                int actualOutput = perceptron.predict(inputs[i]);
                 Assertions.assertEquals(outputs[i], actualOutput, "Failed for input " + Arrays.toString(inputs[i]));
             }
         }
@@ -35,14 +35,14 @@ public class PerceptronTest {
         @Test
         @DisplayName("Perceptron learns 3-input OR gate")
         void testORGateTraining() {
-            double[][] inputs = threeInputTruthTable();
+            float[][] inputs = threeInputTruthTable();
             int[] outputs = {0, 1, 1, 1, 1, 1, 1, 1}; // Only (0, 0, 0) should output 0 for OR gate
 
             Perceptron perceptron = new Perceptron(inputs, outputs, learningRate);
             perceptron.train(100);
 
             for (int i = 0; i < inputs.length; i++) {
-                double actualOutput = perceptron.predict(inputs[i]);
+                int actualOutput = perceptron.predict(inputs[i]);
                 Assertions.assertEquals(outputs[i], actualOutput, "Failed for input " + Arrays.toString(inputs[i]));
             }
         }
@@ -51,9 +51,9 @@ public class PerceptronTest {
         @DisplayName("Perceptron fails to learn XOR gate")
         void testXORGateFailure() {
             // XOR is not linearly separable. The Perceptron MUST fail.
-            double[][] inputs = {
-                    {1.0, 0.0, 0.0}, {1.0, 0.0, 1.0},
-                    {1.0, 1.0, 0.0}, {1.0, 1.0, 1.0}
+            float[][] inputs = {
+                    {1, 0, 0}, {1, 0, 1},
+                    {1, 1, 0}, {1, 1, 1}
             };
             int[] outputs = {0, 1, 1, 0}; // XOR Logic
 
@@ -63,8 +63,8 @@ public class PerceptronTest {
             Assertions.assertFalse(isConverged, "Perceptron should not converge on XOR gate");
         }
 
-        private double[][] threeInputTruthTable() {
-            return new double[][] {
+        private float[][] threeInputTruthTable() {
+            return new float[][] {
                     {1, 0, 0, 0}, {1, 0, 0, 1}, {1, 0, 1, 0}, {1, 0, 1, 1},
                     {1, 1, 0, 0}, {1, 1, 0, 1}, {1, 1, 1, 0}, {1, 1, 1, 1}
             };
@@ -78,22 +78,22 @@ public class PerceptronTest {
         @DisplayName("Learns random linearly separable dataset")
         void testRandomLinearlySeparableData() {
             int trainingSize = 1000;
-            double[][] trainingInputs = new double[trainingSize][3]; // [Bias, X, Y]
+            float[][] trainingInputs = new float[trainingSize][3]; // [Bias, X, Y]
             int[] trainingOutputs = new int[trainingSize];
 
             Random rand = new Random(RANDOM_SEED);
 
             for (int i = 0; i < trainingSize; i++) {
-                trainingInputs[i][0] = 1.0; // Bias input
+                trainingInputs[i][0] = 1; // Bias input
                 // Random x and y between 0 and 100
-                double x = rand.nextDouble() * 100; // 0 to 100
-                double y = rand.nextDouble() * 100; // 0 to 100
+                float x = rand.nextFloat() * 100; // 0 to 100
+                float y = rand.nextFloat() * 100; // 0 to 100
 
                 trainingInputs[i][1] = x;
                 trainingInputs[i][2] = y;
 
                 // Define a linear decision boundary: (2*x + 3*y - 150) >= 0 => output = 1, else output = 0
-                double logicValue = 2 * x + 3 * y - 150;
+                float logicValue = 2 * x + 3 * y - 150;
                 trainingOutputs[i] = logicValue >= 0 ? 1 : 0;
             }
 
@@ -101,17 +101,17 @@ public class PerceptronTest {
             perceptron.train(500);
 
             // Test the trained perceptron on a random input
-            double[][] testInputs = {
-                    {1.0, 10.0, 10.0}, // Should be 0
-                    {1.0, 50.0, 50.0}, // Should be 1
-                    {1.0, 20.0, 30.0}, // Should be 0
-                    {1.0, 5.0, 5.0}, // Should be 0
-                    {1.0, 70.0, 3.0}, // Should be 0 (close to boundary)
-                    {1.0, 0.0, 60.0} // Should be 1
+            float[][] testInputs = {
+                    {1.0f, 10.0f, 10.0f}, // Should be 0
+                    {1.0f, 50.0f, 50.0f}, // Should be 1
+                    {1.0f, 20.0f, 30.0f}, // Should be 0
+                    {1.0f, 5.0f, 5.0f}, // Should be 0
+                    {1.0f, 70.0f, 3.0f}, // Should be 0 (close to boundary)
+                    {1.0f, 0.0f, 60.0f} // Should be 1
             };
             int[] expectedOutputs = {0, 1, 0, 0, 0, 1};
             for (int i = 0; i < testInputs.length; i++) {
-                double predictedOutput = perceptron.predict(testInputs[i]);
+                int predictedOutput = perceptron.predict(testInputs[i]);
                 Assertions.assertEquals(expectedOutputs[i], predictedOutput, "Failed for test input " + Arrays.toString(testInputs[i]));
             }
         }
@@ -119,7 +119,7 @@ public class PerceptronTest {
         @Test
         @DisplayName("Constructor throws exception for mismatched input/output sizes")
         void testMismatchedInputOutputSizes() {
-            double[][] inputs = {
+            float[][] inputs = {
                     {1, 0, 0}, {1, 0, 1},
                     {1, 1, 0}, {1, 1, 1}
             };
@@ -131,7 +131,7 @@ public class PerceptronTest {
         @Test
         @DisplayName("Single sample training")
         void testSingleSample() {
-            double[][] inputs = {{1, 1, 1}};
+            float[][] inputs = {{1, 1, 1}};
             int[] outputs = {1};
 
             Perceptron perceptron = new Perceptron(inputs, outputs, learningRate);
